Replace `ns.sleep` with waiting for done signals.
I kept seeing the amount of RAM available for an iteration change (up and down). I eventually figured out it was due to some of the scripts started lasting just a tiny bit into the next iteration, fouling up the available RAM calculations. Instead of repeatedly checking if scripts are alive, I rely on the final script of each set to send a signal back to the server. This has been far more reliable for immediately knowing when scripts are finished.