monitor.tsx now actually displays some real status.
This adds a new signal SIGNAL_STATUS which monitor.tsx handles by reading the contents of "/run/XXX.txt" where XXX is the pid of the script sending the signal.

Scripts that want status shown on the monitor can use the new `--monitor` flag to know where to send that signal.